+++ lib/ansible/cli/inventory.py
@@ -178,7 +178,7 @@ class InventoryCLI(CLI):
             from ansible.parsing.yaml.dumper import AnsibleDumper
             results = yaml.dump(stuff, Dumper=AnsibleDumper, default_flow_style=False)
         else:
-            from ansible.module_utils.json_utils import jsonify
+            from ansible.module_utils.basic import jsonify
             results = jsonify(stuff, sort_keys=True, indent=4)
 
         return results
